+++ testapp/forms/gallerycollection.py
@@ -2,8 +2,6 @@ from django.forms import fields, widgets
 from django.forms.models import ModelForm
 
 from formset.collection import FormCollection
-from formset.richtext.dialogs import SimpleLinkDialogForm
-from formset.richtext.widgets import RichTextarea, controls
 from formset.widgets import UploadedFileInput
 
 from testapp.models.gallery import Image, Gallery
@@ -14,22 +12,13 @@ class ImageForm(ModelForm):
         required=False,
         widget=widgets.HiddenInput,
     )
-    image = fields.FileField(
-        label="Image",
-        widget=UploadedFileInput,
-        required=False,
-    )
-    caption = fields.CharField(
-        label="Caption",
-        required=False,
-        widget=RichTextarea(
-            #attrs={'use_json': True},
-        )
-    )
 
     class Meta:
         model = Image
         fields = ['id', 'image', 'caption']
+        widgets = {
+            'image': UploadedFileInput,
+        }
 
 
 class ImageCollection(FormCollection):